refactor: separate bearer output from main output
